fixed hardcoded paths and function decomposition
